backend/requirements.txt
Django==5.0.1
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
gunicorn==20.1.0
requests<2.32.0
